requirement.txt
numpy==1.23.5
pandas==2.0.0
SmilesPE==0.03
rdkit-pypi==2022.9.5
scikit-learn==1.2.2
matplotlib==3.7.1
networkx==3.1
tensorboardX==2.6
torch==2.1.2
transformers==4.39.0.dev0
fair-esm==2.0.0
wandb==0.17.0
selfies==2.1.1
peft==0.11.1